chore(web): folder restructure
