+++ web/src/app/layouts/main.njk
@@ -3,7 +3,6 @@
 {% from "govuk/components/back-link/macro.njk" import govukBackLink %}
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 {% from "govuk/components/header/macro.njk" import govukHeader %}
-
 {% block head %}
 	{# setup the header tags, to link to the styles #}
 	<link rel="stylesheet" href="style.css"/>
@@ -14,12 +13,12 @@
 {% endblock %}
 
 {% block header %}
-	{# setup the header tags, to link to the styles #}
 	{{ govukHeader({
 		homepageUrl: "/",
-		serviceName: 'Example To Do App'
+		serviceName: 'Example To Do App',
 	}) }}
 {% endblock %}
+
 
 {% block beforeContent %}
 	{% block backLink %}
