+++ web/src/app/middleware/app.middleware.js
@@ -1,0 +1,7 @@
+export const appMiddleware = (req, res, next) => {
+    const { baseUrl } = req;
+
+    res.locals.baseUrl = baseUrl
+
+    next();
+}
