+++ web/src/app/pages/home/utils/get-page-data-view-model.js
@@ -1,0 +1,3 @@
+
+export const getPageDataViewModel = () =>  ({
+})
