+++ web/src/app/routes/router.js
@@ -1,11 +1,14 @@
 import { Router as createRouter } from 'express';
-import { viewHomepage } from '../home/home.controller.js';
+
+import { viewHomepage } from '../pages/home/home.controller.js';
+import { examplesRouter } from '../pages/examples/router.js';
 
 // create an express router, for handling different paths
 const router = createRouter();
 
-// setup the handler for the homepage: /
+// setup the handlers for the pages
 router.route('/').get(viewHomepage);
+
 
 // export the router for use by the app
 export default router;
