+++ web/src/app/sass/style.scss
@@ -1,0 +1,2 @@
+//Main gov styles
+@import 'node_modules/govuk-frontend/govuk/all';
