--- web/src/app/style.scss
@@ -1,1 +1,0 @@
-@import "govuk-frontend/govuk/all";
