--- web/src/home/home.controller.js
@@ -1,8 +1,0 @@
-/**
- * Render the home page
- *
- * @type {import('express').RequestHandler}
- */
-export function viewHomepage(req, res) {
-    res.render('home/home.view.njk');
-}
